fix update cover bug
